Preparation with ancilla qubit decoherence
This file calculates the lower bound of the output fidelity for the Dicke preparation scheme. Dicke states are prepared using standard phase estimation. And here we consider only ancilla qubit decay and ancilla qubit dephasing.

The lower bound P is the probability that we prepare the target state with 100% fidelity. That is to say, in each round of phase estimation, (1) ancilla qubit decay does not occur in all measurements, (2) the majority vote assigns the correct outcome.